requirements.txt
flask
pillow
gunicorn
tensorflow-cpu
python-dotenv
boto3